+++ exporter/dynatraceexporter/internal/config/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package config // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/dynatraceexporter/config"
+package config // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/dynatraceexporter/internal/config"
 
 import (
 	"errors"
