+++ exporter/dynatraceexporter/internal/serialization/gauge_test.go
